refactor (Frame) : interface -> abstract class로 변경 후 부모에게 중복 로직 위임
