+++ src/main/java/bowling/domain/frame/FinalFrame.java
@@ -1,20 +1,15 @@
 package bowling.domain.frame;
 
-import bowling.domain.frame.state.*;
+import bowling.domain.frame.state.State;
 import bowling.domain.pin.Pins;
 import bowling.domain.score.Score;
 
-import java.util.ArrayList;
-import java.util.Objects;
 import java.util.Optional;
 
-public class FinalFrame implements Frame {
-    private final FrameNumber frameNumber;
-    private States states;
+public class FinalFrame extends Frame {
 
     public FinalFrame(final FrameNumber frameNumber) {
-        this.frameNumber = frameNumber;
-        this.states = new States(new ArrayList<>());
+        super(frameNumber);
     }
 
     @Override
@@ -25,31 +20,13 @@ public class FinalFrame implements Frame {
     }
 
     @Override
-    public State getCurrentState() {
-        if (states.isEmpty() || states.getLast().isTurnOver()) {
-            return new Ready();
-        }
-        return states.getLast();
-    }
-
-    @Override
     public boolean isEnd() {
         return isBonusGameOver() || isGeneralGameOver();
     }
 
     @Override
-    public FrameNumber getFrameNumber() {
-        return frameNumber;
-    }
-
-    @Override
     public Optional<Frame> getNext() {
         return Optional.empty();
-    }
-
-    @Override
-    public States getStates() {
-        return states;
     }
 
     @Override
@@ -76,47 +53,7 @@ public class FinalFrame implements Frame {
         return states.isGeneralGameCount() && states.hasNotBonusState();
     }
 
-    private Score sum() {
-        Score total = Score.INIT_SCORE;
-        for (State state : states.getList()) {
-            total = add(total, state);
-        }
-        return total;
-    }
-
-    private Score add(Score total, final State state) {
-        if (state instanceof Calculable) {
-            Score score = ((Calculable) state).getScore();
-            total = total.add(score);
-        }
-        return total;
-    }
-
     private boolean canAdditionalScore(final Score score) {
         return states.hasCalculableSize(score.getLeftBonusCount());
-    }
-
-    private Score calculate(final Score beforeScore) {
-        Score totalScore = beforeScore;
-        for (State state : states.getList()) {
-            totalScore = totalScore.accumulate(state.getKnockOverCount());
-            if (totalScore.isCompleteAccumulation()) {
-                return totalScore;
-            }
-        }
-        return totalScore;
-    }
-
-    @Override
-    public boolean equals(final Object o) {
-        if (this == o) return true;
-        if (o == null || getClass() != o.getClass()) return false;
-        final FinalFrame that = (FinalFrame) o;
-        return Objects.equals(frameNumber, that.frameNumber);
-    }
-
-    @Override
-    public int hashCode() {
-        return Objects.hash(frameNumber);
     }
 }
