+++ src/main/java/bowling/domain/frame/NormalFrame.java
@@ -1,37 +1,21 @@
 package bowling.domain.frame;
 
-import bowling.domain.frame.state.Calculable;
-import bowling.domain.frame.state.Ready;
 import bowling.domain.frame.state.State;
-import bowling.domain.frame.state.States;
 import bowling.domain.pin.Pins;
 import bowling.domain.score.Score;
 
-import java.util.ArrayList;
-import java.util.Objects;
 import java.util.Optional;
 
-public class NormalFrame implements Frame {
-    private final FrameNumber frameNumber;
+public class NormalFrame extends Frame {
     private final Frame nextFrame;
-    private States states;
 
     public NormalFrame(final FrameNumber frameNumber) {
-        this.frameNumber = frameNumber;
+        super(frameNumber);
         this.nextFrame = createNextFrame();
-        this.states = new States(new ArrayList<>());
     }
 
     public static Frame ofFirst() {
         return new NormalFrame(FrameNumber.ofFirst());
-    }
-
-    private Frame createNextFrame() {
-        final FrameNumber nextFrameNumber = frameNumber.increase();
-        if (nextFrameNumber.isFinal()) {
-            return new FinalFrame(nextFrameNumber);
-        }
-        return new NormalFrame(nextFrameNumber);
     }
 
     @Override
@@ -45,31 +29,13 @@ public class NormalFrame implements Frame {
     }
 
     @Override
-    public State getCurrentState() {
-        if (states.isEmpty() || states.getLast().isTurnOver()) {
-            return new Ready();
-        }
-        return states.getLast();
-    }
-
-    @Override
     public boolean isEnd() {
         return states.getLast().isTurnOver();
     }
 
     @Override
-    public FrameNumber getFrameNumber() {
-        return frameNumber;
-    }
-
-    @Override
     public Optional<Frame> getNext() {
         return Optional.of(nextFrame);
-    }
-
-    @Override
-    public States getStates() {
-        return states;
     }
 
     @Override
@@ -92,42 +58,19 @@ public class NormalFrame implements Frame {
             return Score.NOT_ADDABLE_SCORE;
         }
 
-        Score totalScore = beforeScore;
-        for (State state : states.getList()) {
-            totalScore = totalScore.accumulate(state.getKnockOverCount());
-            if (totalScore.isCompleteAccumulation()) {
-                return totalScore;
-            }
+        Score totalScore = calculate(beforeScore);
+        if (totalScore.isCompleteAccumulation()) {
+            return totalScore;
         }
+
         return nextFrame.calculateAdditionalScore(totalScore);
     }
 
-    private Score sum() {
-        Score total = Score.INIT_SCORE;
-        for (State state : states.getList()) {
-            total = add(total, state);
+    private Frame createNextFrame() {
+        final FrameNumber nextFrameNumber = frameNumber.increase();
+        if (nextFrameNumber.isFinal()) {
+            return new FinalFrame(nextFrameNumber);
         }
-        return total;
-    }
-
-    private Score add(Score total, final State state) {
-        if (state instanceof Calculable) {
-            Score score = ((Calculable) state).getScore();
-            total = total.add(score);
-        }
-        return total;
-    }
-
-    @Override
-    public boolean equals(final Object o) {
-        if (this == o) return true;
-        if (o == null || getClass() != o.getClass()) return false;
-        final NormalFrame normalFrame = (NormalFrame) o;
-        return Objects.equals(frameNumber, normalFrame.frameNumber);
-    }
-
-    @Override
-    public int hashCode() {
-        return Objects.hash(frameNumber);
+        return new NormalFrame(nextFrameNumber);
     }
 }
